doc: update safe navigation view documentation
